refactor(CarNames.java): 자동차이름 생성 로직 분리
Application 클래스에서 객체들을 활용하는 부분이라는걸 명확하게 보여줄 필요를 느껴 수정했습니다
