+++ src/main/java/racingcar/Application.java
@@ -2,9 +2,9 @@ package racingcar;
 
 import camp.nextstep.edu.missionutils.Console;
 
-import java.util.Arrays;
 import java.util.List;
 
+import static racingcar.CarNames.createCarNames;
 import static racingcar.WinnerCalculator.calculateWinners;
 
 public class Application {
@@ -26,18 +26,5 @@ public class Application {
         List<String> winnerNames = calculateWinners(lastRacingRecord);
         String winnerRecord = String.join(", ", winnerNames);
         System.out.println("최종 우승자 : " + winnerRecord);
-    }
-
-    public static List<String> createCarNames(String carName) {
-        List<String> carNames = Arrays.stream(carName.split(","))
-                .toList();
-
-        for (String name : carNames) {
-            if (name.length() > 5) {
-                throw new IllegalArgumentException(name + "이름이 5자리가 넘습니다");
-            }
-        }
-
-        return carNames;
     }
 }
